+++ dev/pages/settings/theme-page.tsx
@@ -1,9 +1,4 @@
 import React from 'react';
 import { ThemeEditor } from 'eidos-ui';
-import { SettingsLayout } from './settings-layout';
 
-export const ThemePage: React.FC = () => (
-  <SettingsLayout>
-    <ThemeEditor />
-  </SettingsLayout>
-);
+export const ThemePage: React.FC = () => <ThemeEditor />;